requirements.txt
bson==0.5.10
click==7.1.2
dnspython==2.1.0
fastapi==0.63.0
gunicorn==20.0.4
h11==0.12.0
install==1.3.4
pydantic==1.7.3
pymongo==3.11.2
python-dateutil==2.8.1
python-decouple==3.4
six==1.15.0
starlette==0.13.6
uvicorn==0.13.3
